# Optimised version of fused classifier + bugfixes(?)